fix: update ImageFile root and table work again
