--- src/Component/ImageFile/ImageFileEditForm/ImageFileEditForm.less
@@ -1,7 +1,0 @@
-.imagefile-form {
-  .form-buttons {
-    button.ant-btn {
-      margin: 0 5px;
-    }
-  }
-}
