+++ src/Component/ImageFile/ImageFileRoot/ImageFileRoot.less
@@ -1,14 +1,15 @@
-.imagefile-root{
+.imagefile-root {
   height: 100%;
   display: grid;
   grid-template-rows: auto 1fr;
-  grid-template-columns: 40% 60%;
+  grid-template-columns: 50% 50%;
   grid-template-areas:
     "header header"
     "left right";
 
   >div {
     flex: 1;
+    overflow: auto;
 
     &.header {
       grid-area: header;
@@ -18,14 +19,14 @@
       grid-area: left;
       padding: 0 10px 0 24px;
 
-      background:
-        linear-gradient(@primary-color, @primary-color) right / 2px 90%,
-      rgba(255, 255, 255, 0);
-      background-repeat:no-repeat;
+      .left-toolbar {
+        padding: 0 0 10px 0;
+      }
     }
 
     &.right-container {
       grid-area: right;
+      border-left: solid 2px @border-color-base;
       padding: 0 24px 0 10px;
     }
   }
