+++ src/Component/ImageFile/ImageFileTable/ImageFileTable.less
@@ -16,6 +16,8 @@
 
     .operation-column {
       font-size: 1.5em;
+      text-align: center;
+      color: @primary-color;
 
       .actions {
         display: flex;
