--- src/Component/Table/EntityTable/EntityTable.less
@@ -1,41 +1,0 @@
-.entity-table {
-
-  .ant-table {
-    .ant-table-selection-column {
-      width: 0;
-      padding: 0;
-    }
-
-    .ant-table-row {
-      cursor: pointer;
-
-      &:hover {
-        color: @primary-color;
-      }
-    }
-
-    th.operation-column {
-      font-size: 1.5em;
-      text-align: center;
-      color: @primary-color;
-    }
-
-    .operation-column {
-      font-size: 1.5em;
-
-      .actions {
-        display: flex;
-        justify-content: space-around;
-
-        .anticon {
-          cursor: pointer;
-
-          &.anticon-delete {
-            color: #ea2f5a;
-          }
-        }
-      }
-    }
-  }
-
-}
